Add files via upload
In this project I used data from the BigQuery website and I had to use it to find trend in the homelessness population. I came up with queries to put the data in ascending and descending order. To find out which cities have more and less unsheltered homeless people. Also used a chart to answer what states are over representing and underrepresented against a Top 10 chart in regards to the overall homeless population.